subtask(F-4-5-split-my-permissions-web-view-model): Split my-permissions page state from rendering
